Update tests script. Add a script to quickly run specific tests.
